benford.py
import pandas as pd

def benford_law(df):
    """
    This function calculates the distribution of the first digit of each number in a given dataframe.

    Parameters:
    df (DataFrame): The input dataframe.

    Returns:
    list: A list of dictionaries, each containing a digit and its frequency.
    """
    # Extract the first digit of each number
    first_digits = [int(str(num)[0]) for num in df.values.flatten() if str(num)[0].isdigit()]

    # Count the occurrences of each digit
    digit_counts = pd.Series(first_digits).value_counts().sort_index()

    # Calculate the distribution
    distribution = digit_counts / digit_counts.sum()

    # Convert the distribution to a list of dictionaries
    distribution = [{'digit': str(digit), 'frequency': round(frequency, 3)} for digit, frequency in distribution.items()]

    return distribution

def follows_benford_law(results):
    """
    This function checks if a given distribution follows Benford's Law.

    Parameters:
    results (list): The input distribution, a list of dictionaries each containing a digit and its frequency.

    Returns:
    bool: True if the distribution follows Benford's Law, False otherwise.
    """
    benford_frequencies = [0.301, 0.176, 0.125, 0.097, 0.079, 0.067, 0.058, 0.051, 0.046]
    actual_frequencies = [result['frequency'] for result in results]
    differences = [abs(a - b) for a, b in zip(actual_frequencies, benford_frequencies)]
    return all(difference <= 0.02 for difference in differences)  # Adjust the tolerance as needed